Move save and restore operation to a dedicated command
Save and restore are required for any commands and not only for the
USER_DATA.
Two new commands are created for that and README is changed accordingly.
For these new commands BMT456 does not exepct call to sendbyte but a
write byte on the command with a dummy value.
A test is done on the BMR version to know how to execute the command.

Moreover for the restore two options are provided to reload the default
configuration or the last saved configuration.
No automatic save is done after a restore.